Update workflows.md
Moving away from legal terms definition, for which it is not the right place, to a technical terms glossary.
Removing: Contracting Party, PAF Ecosystem.

Removing Transaction as "the sending of PAF Data from the Root Party through the PAF ecosystem by consecutive Transmissions". This does not seem to be used anywhere else in the documentation, and it's confusing with "Transaction ID", which is an ad slot bid request identifier as described in mvp-spec/json-schemas/transaction_id.json 

Removed Pseudo-Identifier definition, as describing the type of identifier in "User Id and Preferences" is enough.